Small refactoring and added SeceurityConfig for authApi
