+++ authApi/src/main/java/cs4337/group9/authapi/AuthApiApplication.java
@@ -1,4 +1,4 @@
-package cs4337.group9.contentapi;
+package cs4337.group9.authapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,12 +9,10 @@ import java.util.Map;
 
 @SpringBootApplication
 @RestController
-public class ContentApiApplication {
+public class AuthApiApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ContentApiApplication.class, args);
+        SpringApplication.run(AuthApiApplication.class, args);
     }
 
-    @GetMapping( "/hello" ) public Map<String, Object> sayHello () {
-        return Map.of( "message" , "Hello World!" ); }
 }
